requirements.txt
paho-mqtt==1.6.1
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
redbird==0.6.1
rocketry==2.5.1
six==1.16.0
typing_extensions==4.4.0
yapf==0.32.0
